INT-1406: Install fix for when not upgrading from 1.9
git-svn-id: http://svn.moodlerooms.com/Moodle/2.0/branches/plugins@15805 f529b7b8-409e-e885-f664-9768b4ed42ab

+++ db/install.php
@@ -457,7 +457,9 @@ function xmldb_format_flexpage_install() {
     events_uninstall('block/page_module');
 
 /// Cleanup mod/pagemenu
-    uninstall_plugin('mod', 'pagemenu');
+    if ($DB->record_exists('modules', array('name' => 'pagemenu'))) {
+        uninstall_plugin('mod', 'pagemenu');
+    }
 
 /// Cleanup course/format/page
     foreach (array('format_page', 'format_page_items') as $table) {
